requirements.txt
langgraph 
langsmith 
langchain
langchain-community 
langchain-openai 
tavily-python 
langgraph-cli[inmem] 
langchain-ollama
python-dotenv
langchain 
chromadb 
sentence-transformers 
transformers 
pandas 
langchain_chroma